some docs, and move FETCH_OPTIONS to the fetch api
