+++ src/scope/network/fs/fs.ts
@@ -1,6 +1,6 @@
 import { deprecate, fetch, put, remove, undeprecate } from '../../../api/scope';
+import { FETCH_OPTIONS } from '../../../api/scope/lib/fetch';
 import { BitId } from '../../../bit-id';
-import { FETCH_OPTIONS } from '../../../constants';
 import ComponentsList, { ListScopeResult } from '../../../consumer/component/components-list';
 import Component from '../../../consumer/component/consumer-component';
 import ComponentObjects from '../../component-objects';
